backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
celery==5.3.6
langchain-openai==0.0.2
pydantic-settings==2.1.0
reportlab==4.0.9